CMDB::YAML: Expand YAML references in CMDB pathnames
If a pathname in the list of pathnames given as an argument to `set
cmdb` contains a reference, this will be expanded using the current set
of parsed YAML data which allows flexibility.

A reference is given in square brackets and it will be replaced with the
corresponding value in case the value is a scalar (a value not a key in
YAML speak) or an array of scalars.

E.g:

Given,
```
set cmdb => {
	type => YAML,
	path => [
		'cmdb/{hostname}.yml',
		'cmdb/defaults/[machine.defaults].yml'
	]
}
```

Sex will look for the keys `machine` & `defaults` in
`cmdb/{hostname}.yml`, replace the value in the pathname with the
corresponding value (or leave it intact if the value is non-existent or
is not a value but a key, ie, points to another variable) and then move
on to parse this pathname.

YAML arrays work in a similar way, given

```
my:
  array:
    - foo
    - bar
    - baz
```

in `cmdb/{hostname}.yml` and given the cmdb,

```
set cmdb => {
	type => YAML,
	path => [
		'cmdb/{hostname}.yml',
		'cmdb/roles/[my.array].yml'
	]
}
```

Sex will try & include the files `cmdb/roles/{foo,bar,baz}.yml` in the
given order.

Expansion happens during parsing using the data so-far parsed in the
list of paths.