Adds SLE 15 / Leap 15.0 support
Replaced "SUSE Linux Enterprise Server" with "SUSE Linux Enterprise". It does not matter, if it is a server or desktop installation. It just does matter, that it is an enterprise installation.
Preferred to use "/etc/os-release" file (available since SLE 11.4) instead of deprecated "/etc/SuSE-release", that is not available anymore.